refactor: modularize player zone widget
